Add SHA1 to [GET] /api/file response
Hey!

I was noticing that SHA1 was missing from the API response while searching for files using [GET] /api/file and using queries for searching.
While it doesn't make sense to include every type of hash in the API response, I'd argue that SHA1 is still in common use. It would reduce to the amount of API calls to fetch SHA1 hashes en-masse. 

Please let me know if it's not feasible or if there is an alternative method recommended. :)